requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pyjwt==2.9.0
firebase-admin==6.5.0
psycopg2-binary==2.9.9
pymysql==1.1.1
cryptography==42.0.5
sqlalchemy==2.0.35
python-multipart==0.0.9
bcrypt==4.1.2
email-validator==2.1.0
